+++ lib/rpmts.h
@@ -775,7 +775,7 @@ void rpmtsCheckDSIProblems(const rpmts ts, const rpmte te)
  */
 /*@null@*/
 void * rpmtsNotify(rpmts ts, rpmte te,
-                rpmCallbackType what, unsigned long long amount, unsigned long long total)
+                rpmCallbackType what, uint64_t amount, uint64_t total)
 	/*@*/;
 
 /**
